refactor: relocate audit subsystem into wyrelog/audit subdirectory
The flat-prefix directory rule maps each wyl_<domain>_* prefix
onto its own subdirectory of wyrelog/. Audit was previously laid
out as wyrelog/wyl-audit-*.{c,h}, doubling the audit token in
both filename and function symbol. Move the four files that make
up the wyl_audit_* prefix into wyrelog/audit/, dropping the
redundant wyl-audit- prefix from filenames:

  wyrelog/wyl-audit-event.c          -> wyrelog/audit/event.c
  wyrelog/wyl-audit-conn.c           -> wyrelog/audit/conn.c
  wyrelog/wyl-audit-conn-private.h   -> wyrelog/audit/conn-private.h
  wyrelog/wyl-audit-iter.c           -> wyrelog/audit/iter.c

The public umbrella header wyrelog/audit.h stays at the top level
because it is part of the installed include set and the umbrella
shape (one .h per domain alongside one wyrelog.h) is the public-
facing convention.

Function and type names are unchanged; only file paths move. The
single internal include in audit/conn.c switches from
"wyl-audit-conn-private.h" to the sibling-relative
"conn-private.h", and tests/test-audit-conn.c follows the public
include path "wyrelog/audit/conn-private.h".

Verified locally:
  - meson setup builddir (default, audit=disabled): 23/23 PASS,
    no behavioural change.
  - meson setup builddir-audit -Denable_audit=enabled: 24/24
    PASS including audit-event and audit-conn.

Other prefixes (session, decide, perm, handle, ...) keep the flat
layout for now; their migration into per-prefix subdirectories
can land as separate atomics if and when their per-domain file
counts justify it.
